--- conexao/services/Constantes.js
@@ -1,3 +1,0 @@
-export class Constantes {
-    static urlBase = 'http://localhost8080/v1';
-}
